Extract feedback_builders.rs from turn.rs (parent #680)
Create new sibling module `src/agent/loop_run/feedback_builders.rs`
to host the FeedbackFrame builder helpers + path-extraction utilities
used by the bash / edit failure pipelines. Small follow-on extraction
in the meta-issue #680 reduction effort — the helpers form a clean
responsibility-focused cluster.

Migrated to `feedback_builders.rs` (`pub(super)`):

* `build_feedback_for_bash(&BashExecutionOutcome, &Path)
    -> Option<FeedbackFrame>`
* `bash_outcome_primary_error(&BashExecutionOutcome) -> Option<String>`
* `build_feedback_for_unsafe_block_reason(command, rendered_reason, &Path)
    -> FeedbackFrame` (Issue #461 / DR4-004: rendered reason in
    primary_error, raw command in command field only)
* `build_feedback_for_edit_failure(path, err, &Path) -> FeedbackFrame`
    (CB-001)
* `extract_suspected_files_from_text(stdout, stderr) -> Vec<PathBuf>`
    (best-effort cap-8 heuristic)
* `extract_path_tokens_from_text(text, &Path) -> Vec<String>`
* `extract_current_request_paths(&Agent, &Path) -> Vec<String>`
    (caps to `prompting::MAX_CURRENT_REQUEST_PATHS`)

`loop_run.rs` adjustments:

* Registered `mod feedback_builders;` with a parent-#680 doc comment.

`turn.rs` adjustments:

* Removed the 7 fn definitions (originally at lines 690-851).
* Added the 5 caller-facing imports
  (`build_feedback_for_{bash,edit_failure,unsafe_block_reason}` /
  `extract_current_request_paths` /
  `extract_suspected_files_from_text`) via a new
  `use super::feedback_builders::{...}` block.
* Removed the now-unused `extract_path_like_tokens` import from the
  existing `super::verifier_repair_targeting::{...}` block (the
  symbol is still used by `feedback_builders` via the same module
  path).

LOC delta:

* `turn.rs`: 28,734 → 28,557 (-177 LOC)
* `feedback_builders.rs`: 0 → 200 (new)
* Cumulative `turn.rs`: 39,489 → 28,557 (-10,932 LOC, -27.7%)

`pub(super)` limited / no facade re-export (DR3-001).
`turn.rs` remains the only in-crate consumer.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
